Rename `Navigator` component to `NavigatorRoot`, move code under `navigator-root` folder
