+++ packages/components/src/navigator/index.js
@@ -1,3 +1,3 @@
-export { default as Navigator } from './navigator';
+export { default as NavigatorRoot } from './navigator-root';
 export { default as NavigatorScreen } from './screen';
 export { default as useNavigator } from './use-navigator';
