+++ packages/components/src/navigator/navigator-root/README.md
@@ -1,18 +1,18 @@
-# Navigator
+# `NavigatorRoot`
 
 <div class="callout callout-alert">
 This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
 </div>
 
-The Navigator components allows rendering nested panels or menus (also called screens) and navigate between these different states. The Global Styles sidebar is an example of this.
+The `NavigatorRoot` component allows rendering nested panels or menus (via the [`NavigatorScreen` component](/packages/components/src/navigator/navigator-screen/README.md)) and navigate between these different states (via the `useNavigator` hook). The Global Styles sidebar is an example of this.
 
-The components is not opinionated in terms of UI, it lets compose any UI components to navigate between the nested screens.
+The `Navigator*` family of components is _not_ opinionated in terms of UI, and can be composed with any UI components to navigate between the nested screens.
 
 ## Usage
 
 ```jsx
 import {
-	__experimentalNavigator as Navigator,
+	__experimentalNavigatorRoot as NavigatorRoot,
 	__experimentalNavigatorScreen as NavigatorScreen,
 	__experimentalUseNavigator as useNavigator,
 } from '@wordpress/components';
@@ -32,7 +32,7 @@ function NavigatorButton( {
 }
 
 const MyNavigation = () => (
-	<Navigator initialPath="/">
+	<NavigatorRoot initialPath="/">
 		<NavigatorScreen path="/">
 			<p>This is the home screen.</p>
 			<NavigatorButton isPrimary path="/child">
@@ -46,43 +46,31 @@ const MyNavigation = () => (
 				Go back
 			</NavigatorButton>
 		</NavigatorScreen>
-	</Navigator>
+	</NavigatorRoot>
 );
 ```
 
-## Navigator Props
+## Props
 
-`Navigator` supports the following props.
+The component accepts the following props:
 
-### `initialPath`
-
--   Type: `string`
--   Required: No
+### `initialPath`: `string`
 
 The initial active path.
 
-## NavigatorScreen Props
+-   Required: No
 
-`NavigatorScreen` supports the following props.
 
-### `path`
-
--   Type: `string`
--   Required: Yes
-
-The path of the current screen.
-
-## The navigator object.
+## The navigator object
 
 You can retrieve a `navigator` instance by using the `useNavigator` hook.
-The navigator offers the following methods:
 
-### `push`
+The hook offers the following methods:
 
--   Type: `( path: string, options ) => void`
+### `push`: `( path: string, options ) => void`
 
 The `push` function allows you to navigate to a given path. The second argument can augment the navigation operations with different options.
 
 The available options are:
 
- - `isBack` (`boolean): A boolean flag indicating that we're moving back to a previous state.
+- `isBack` (`boolean): A boolean flag indicating that we're moving back to a previous state. -->
