+++ packages/components/src/navigator/navigator-root/index.js
@@ -1,0 +1,1 @@
+export { default } from './component';
